backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
langchain-core>=0.3.31
langchain-anthropic>=0.3.4
langchain-openai>=0.3.0
langchain-google-genai>=2.1.0
langchain-ollama>=0.3.0
langgraph>=0.2.60
chromadb>=0.6.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.2
sentence-transformers>=3.3.0
websockets>=14.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
# Phase 2: Persistence
sqlmodel>=0.0.22
psycopg2-binary>=2.9.0
langgraph-checkpoint-postgres>=2.0.0
alembic>=1.14.0
# Phase 3: Hybrid search
rank-bm25>=0.2.2
# Phase 4: Scalability
redis>=5.0.0
arq>=0.26.0
# Phase 5: Evaluation
ragas>=0.2.0
# Phase 6: Coding tools
docker>=7.0.0
